backend/requirements.txt
flask
flask-cors
python-dotenv
groq
pyautogui
mss
Pillow
sounddevice
soundfile
openai